Refactored how cookie domain works. Now consistent with request plugin
