+++ plugins/insomnia-plugin-cookie-jar/__tests__/index.js
