+++ plugins/insomnia-plugin-cookie-jar/package.json
